+++ scripts/sql_s32_batch.sql
@@ -1,0 +1,91 @@
+BEGIN;
+
+-- Comancheria (id=216) - 1700-1875
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(216, 'Hämäläinen, P. (2008). The Comanche Empire. Yale University Press. ISBN 978-0300126549.', 'https://yalebooks.yale.edu/9780300126549', 'academic'),
+(216, 'Kavanagh, T. W. (1996). Comanche Political History: An Ethnohistorical Perspective, 1706-1875. University of Nebraska Press. ISBN 978-0803227293.', 'https://www.nebraskapress.unl.edu/nebraska/9780803227293/', 'academic'),
+(216, 'Wallace, E. & Hoebel, E. A. (1952). The Comanches: Lords of the South Plains. University of Oklahoma Press. ISBN 978-0806113944.', 'https://www.oupress.com/9780806113944/the-comanches/', 'academic'),
+(216, 'Brooks, J. F. (2002). Captives and Cousins: Slavery, Kinship, and Community in the Southwest Borderlands. University of North Carolina Press. ISBN 978-0807853825.', 'https://uncpress.org/book/9780807853825/captives-and-cousins/', 'academic'),
+(216, 'DeLay, B. (2008). War of a Thousand Deserts: Indian Raids and the U.S.-Mexican War. Yale University Press. ISBN 978-0300119329.', 'https://yalebooks.yale.edu/9780300119329', 'academic');
+
+-- Lakota Oyate (id=217) - 1600-1877
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(217, 'Hämäläinen, P. (2019). Lakota America: A New History of Indigenous Power. Yale University Press. ISBN 978-0300215953.', 'https://yalebooks.yale.edu/9780300215953', 'academic'),
+(217, 'Andersson, R.-H. & Posthumus, D. C. (2022). Lakȟóta: An Indigenous History. University of Oklahoma Press. ISBN 978-0806191904.', 'https://www.oupress.com/9780806191904/lakhota/', 'academic'),
+(217, 'DeMallie, R. J. (2001). Teton. In: R. J. DeMallie (ed.), Handbook of North American Indians, Vol. 13: Plains, Part 2, pp. 794-820. Smithsonian Institution. ISBN 978-0160504013.', 'https://www.bookstore.gpo.gov', 'academic'),
+(217, 'Ostler, J. (2010). The Lakotas and the Black Hills: The Struggle for Sacred Ground. Penguin/Viking. ISBN 978-0143119753.', 'https://www.penguinrandomhouse.com', 'academic'),
+(217, 'Powers, W. K. (1975). Oglala Religion. University of Nebraska Press. ISBN 978-0803287068.', 'https://www.nebraskapress.unl.edu/nebraska/9780803287068/', 'academic');
+
+-- Mvskoke / Creek Confederacy (id=219) - 1600-1832
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(219, 'Saunt, C. (1999). A New Order of Things: Property, Power, and the Transformation of the Creek Indians, 1733-1816. Cambridge University Press. ISBN 978-0521660433.', 'https://www.cambridge.org/9780521660433', 'academic'),
+(219, 'Martin, J. W. (1991). Sacred Revolt: The Muskogees'' Struggle for a New World. Beacon Press. ISBN 978-0807054192.', 'https://www.beacon.org/Sacred-Revolt-P190.aspx', 'academic'),
+(219, 'Ethridge, R. F. (2003). Creek Country: The Creek Indians and Their World. University of North Carolina Press. ISBN 978-0807853962.', 'https://uncpress.org/book/9780807853962/creek-country/', 'academic'),
+(219, 'Wright, J. L. Jr. (1990). The Only Land They Knew: The Tragic Story of the American Indians in the Old South. University of Nebraska Press. ISBN 978-0029354506.', 'https://www.nebraskapress.unl.edu/nebraska/9780029354506/', 'academic'),
+(219, 'Hahn, S. C. (2004). The Invention of the Creek Nation, 1670-1763. University of Nebraska Press. ISBN 978-0803273856.', 'https://www.nebraskapress.unl.edu/nebraska/9780803273856/', 'academic');
+
+-- Tundo had id=367 but I notice this was already done in S31. Let me skip ID 367.
+-- Actually it was done in S31. Need different entity.
+
+-- Madja-as (id=884) - 1212-1569 Panay Philippines
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(884, 'Scott, W. H. (1984). Prehispanic Source Materials for the Study of Philippine History (rev. ed.). New Day Publishers. ISBN 978-9711002268.', 'https://www.worldcat.org/title/prehispanic-source-materials/oclc/12030547', 'academic'),
+(884, 'Scott, W. H. (1994). Barangay: Sixteenth-Century Philippine Culture and Society. Ateneo de Manila University Press. ISBN 978-9715501354.', 'https://www.ateneo.edu/aupress', 'academic'),
+(884, 'Monteclaro, P. A. (1907). Maragtas (Hiligaynon/Kinaray-a). Iloilo. (Repr. with Eng. transl. in J. R. Hosillos, Originality as Vengeance in Philippine Literature, 1984)', 'https://www.worldcat.org/title/maragtas/oclc/13550028', 'academic'),
+(884, 'Beyer, H. O. (1949). Outline Review of Philippine Archaeology by Islands and Provinces. Philippine Journal of Science, 77(3-4), 205-374.', 'https://www.science.gov.ph', 'academic'),
+(884, 'Talaguit, C. J. N. (2019). Mga Maragtas ng Panay: Comparative Analysis of Documents about the Bornean Settlement Tradition. De La Salle University thesis.', 'https://animorepository.dlsu.edu.ph', 'academic');
+
+-- Cotabato/Buayan (id=898) - 1350-1520 Mindanao pre-Maguindanao
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(898, 'Laarhoven, R. (1986). We Are Many Nations: The Emergence of a Multi-Ethnic Maguindanao Sultanate. Philippine Quarterly of Culture and Society, 14(1), 32-53.', 'https://www.jstor.org/stable/29791929', 'academic'),
+(898, 'Hayase, S. (2007). Mindanao Ethnohistory Beyond Nations: Maguindanao, Sangir, and Bagobo Societies. Ateneo de Manila University Press. ISBN 978-9715505116.', 'https://www.ateneo.edu/aupress', 'academic'),
+(898, 'Donoso, I. (2023). Bichara: Moro Chanceries and Jawi Legacy in the Philippines. Springer. ISBN 978-9811908200.', 'https://link.springer.com/book/10.1007/978-981-19-0821-7', 'academic'),
+(898, 'Saleeby, N. M. (1905). Studies in Moro History, Law, and Religion. Bureau of Public Printing, Manila. (Repr. Cellar Book Shop)', 'https://www.worldcat.org/title/studies-in-moro-history/oclc/2300064', 'academic'),
+(898, 'Majul, C. A. (1973). Muslims in the Philippines (2nd ed.). University of the Philippines Press. ISBN 978-9711050017.', 'https://up.edu.ph/', 'academic');
+
+-- Gangga Negara (id=902) - 100-1025 Malay Peninsula
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(902, 'Cœdès, G. (1968). The Indianized States of Southeast Asia. University of Hawai''i Press. ISBN 978-0824803681.', 'https://uhpress.hawaii.edu/title/the-indianized-states-of-southeast-asia/', 'academic'),
+(902, 'Brown, C. C. (transl.) (1952). Sejarah Melayu / Malay Annals. Journal of the Malayan Branch of the Royal Asiatic Society, 25(2-3), 5-276.', 'https://www.jstor.org/stable/41502950', 'academic'),
+(902, 'Othman bin Mohd. Yatim et al. (1994). Beruas: Kerajaan Melayu Kuno di Perak. Persatuan Sejarah Malaysia. ISBN 978-9839610017.', 'https://www.sejarah.com.my', 'academic'),
+(902, 'Jacq-Hergoualc''h, M. (2002). The Malay Peninsula: Crossroads of the Maritime Silk Road (100 BC-AD 1300). Brill. ISBN 978-9004119734.', 'https://brill.com/view/title/8126', 'academic'),
+(902, 'Miksic, J. N. & Goh, G. Y. (2017). Ancient Southeast Asia. Routledge. ISBN 978-0415735544.', 'https://www.routledge.com/9780415735544', 'academic');
+
+-- Gelgel Bali (id=906) - 1343-1550
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(906, 'Hägerdal, H. (2001). Hindu Rulers, Muslim Subjects: Lombok and Bali in the Seventeenth and Eighteenth Centuries. White Lotus Press. ISBN 978-9747534221.', 'https://www.whitelotuspress.com', 'academic'),
+(906, 'Robson, S. O. (1995). Deśawarṇana (Nāgarakṛtāgama) by Mpu Prapañca. KITLV Press. ISBN 978-9067180641.', 'https://brill.com/view/title/22898', 'academic'),
+(906, 'Vickers, A. (1989). Bali: A Paradise Created. Periplus Editions. ISBN 978-9625930633.', 'https://www.tuttlepublishing.com', 'academic'),
+(906, 'Schulte Nordholt, H. (1996). The Spell of Power: A History of Balinese Politics, 1650-1940. KITLV Press. ISBN 978-9067181426.', 'https://brill.com/view/title/22906', 'academic'),
+(906, 'Reuter, T. (2002). Custodians of the Sacred Mountains: Culture and Society in the Highlands of Bali. University of Hawai''i Press. ISBN 978-0824824501.', 'https://uhpress.hawaii.edu/title/custodians-of-the-sacred-mountains/', 'academic');
+
+-- Dyolof pre-1549 (id=984) - 1200-1549 Wolof
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(984, 'Boulègue, J. (2013). Les royaumes wolof dans l''espace sénégambien (XIIIe-XVIIIe siècle). Karthala Editions. ISBN 978-2811108816.', 'https://www.karthala.com', 'academic'),
+(984, 'Barry, B. (1998). Senegambia and the Atlantic Slave Trade. Cambridge University Press. ISBN 978-0521597265.', 'https://www.cambridge.org/9780521597265', 'academic'),
+(984, 'Charles, E. A. (1977). Precolonial Senegal: The Jolof Kingdom, 1800-1890. African Studies Center, Boston University.', 'https://www.bu.edu/africa', 'academic'),
+(984, 'Diouf, M. (2001). Histoire du Sénégal: Le modèle islamo-wolof et ses périphéries. Maisonneuve & Larose. ISBN 978-2706815034.', 'https://www.cnrseditions.fr', 'academic'),
+(984, 'Levtzion, N. & Pouwels, R. L. (eds.) (2000). The History of Islam in Africa. Ohio University Press. ISBN 978-0821412961.', 'https://www.ohioswallow.com/book/The+History+of+Islam+in+Africa', 'academic');
+
+-- Mogadishu pre-Fakhr ad-Din (id=1009) - 900-1270
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(1009, 'Jama, A. (1996). The Origins and Development of Mogadishu AD 1000 to 1850. Studies in African Archaeology 12. Uppsala University. ISBN 978-9150611236.', 'https://www.diva-portal.org/smash/record.jsf?pid=diva2%3A167611', 'academic'),
+(1009, 'Mukhtar, M. H. (2003). Historical Dictionary of Somalia (New ed.). Scarecrow Press. ISBN 978-0810866041.', 'https://rowman.com/ISBN/9780810866041', 'academic'),
+(1009, 'Cassanelli, L. V. (1982). The Shaping of Somali Society: Reconstructing the History of a Pastoral People, 1600-1900. University of Pennsylvania Press. ISBN 978-0812278323.', 'https://www.upenn.edu/pennpress/book/13888.html', 'academic'),
+(1009, 'Chittick, H. N. (1982). Medieval Mogadishu. Paideuma, 28, 45-62.', 'https://www.jstor.org/stable/40341670', 'academic'),
+(1009, 'Pouwels, R. L. (1987). Horn and Crescent: Cultural Change and Traditional Islam on the East African Coast, 800-1900. Cambridge University Press. ISBN 978-0521523097.', 'https://www.cambridge.org/9780521523097', 'academic');
+
+-- Milagro-Quevedo / Manteño (id=968) - 500-1535 coastal Ecuador
+INSERT INTO sources (entity_id, citation, url, source_type) VALUES
+(968, 'Marcos, J. G. (1981). Real Alto: La historia de un centro ceremonial Valdivia, 2 vols. Biblioteca Ecuatoriana de Arqueología. Escuela Superior Politécnica del Litoral.', 'https://www.espol.edu.ec', 'academic'),
+(968, 'Estrada, V. E. (1962). Arqueología de Manabí Central. Museo Víctor Emilio Estrada. ISBN 978-9978920038.', 'https://www.worldcat.org/title/arqueologia-de-manabi-central/oclc/1755066', 'academic'),
+(968, 'Mester, A. M. (1990). The Pearl Divers of Los Frailes: Archaeological and Ethnohistorical Explorations of Sumptuary Goods Trade and Cosmology in the North and Central Andes. PhD diss., University of Illinois Urbana-Champaign.', 'https://www.ideals.illinois.edu', 'academic'),
+(968, 'McEwan, C. & Hudson, C. (1994). Power and Prestige in Pre-Hispanic Manabí. In: A. M. Maúgüé (ed.), Antes del Estado Inca. Banco Central del Ecuador. ISBN 978-9978920458.', 'https://www.bce.fin.ec', 'academic'),
+(968, 'Bray, T. L. (ed.) (2008). The Archaeology of Wak''as: Explorations of the Sacred in the Pre-Columbian Andes. University Press of Colorado. ISBN 978-1607323471.', 'https://upcolorado.com/university-press-of-colorado/item/2536-the-archaeology-of-wakas', 'academic');
+
+-- Confidence score upgrades
+UPDATE geo_entities SET confidence_score = 0.85 WHERE id IN (216, 217, 219, 884, 898, 902, 906, 984, 1009, 968);
+
+COMMIT;
+
+SELECT g.id, g.name_original, g.confidence_score, (SELECT count(*) FROM sources s WHERE s.entity_id=g.id) AS n_src FROM geo_entities g WHERE g.id IN (216, 217, 219, 884, 898, 902, 906, 968, 984, 1009) ORDER BY g.id;